tools/opai-studio/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.25
python-jose[cryptography]>=3.3.0
pillow>=10.0
aiofiles>=23.2.0
